Bugfix: correct reference to CC.
