+++ tests/qat/serialised_compiler_config_templates/serialised_full_compiler_config_v02.json
@@ -1,1 +1,1 @@
-{"$type": "<class 'compiler_config.CompilerConfig'>", "$data": {"repeats": 1000, "repetition_period": 10, "results_format": {"$type": "<class 'scc.compiler.config.QuantumResultsFormat'>", "$data": {"format": {"$type": "<enum 'scc.compiler.config.InlineResultsProcessing'>", "$value": 2}, "transforms": {"$type": "<enum 'scc.compiler.config.ResultsFormatting'>", "$value": 2}}}, "metrics": {"$type": "<enum 'scc.compiler.config.MetricsType'>", "$value": 4}, "active_calibrations": [], "optimizations": {"$type": "<class 'scc.compiler.config.Qasm2Optimizations'>", "$data": {"qiskit_optimizations": {"$type": "<enum 'scc.compiler.config.QiskitOptimizations'>", "$value": 1}, "tket_optimizations": {"$type": "<enum 'scc.compiler.config.TketOptimizations'>", "$value": 18}}}}}
+{"$type": "<class 'compiler_config.config.CompilerConfig'>", "$data": {"repeats": 1000, "repetition_period": 10, "results_format": {"$type": "<class 'compiler_config.config.QuantumResultsFormat'>", "$data": {"format": {"$type": "<enum 'compiler_config.config.InlineResultsProcessing'>", "$value": 2}, "transforms": {"$type": "<enum 'compiler_config.config.ResultsFormatting'>", "$value": 2}}}, "metrics": {"$type": "<enum 'compiler_config.config.MetricsType'>", "$value": 4}, "active_calibrations": [], "optimizations": {"$type": "<class 'compiler_config.config.Qasm2Optimizations'>", "$data": {"qiskit_optimizations": {"$type": "<enum 'compiler_config.config.QiskitOptimizations'>", "$value": 1}, "tket_optimizations": {"$type": "<enum 'compiler_config.config.TketOptimizations'>", "$value": 18}}}}}
